@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.17.6 <3.0.0"
+  sdk: ">=3.0.3 <3.10.4"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -34,6 +34,8 @@ dependencies:
   analyzer: ^3.3.1
   dio: ^4.0.4
   json_serializable: ^6.1.4
+  dartz: ^0.10.1
+  internet_connection_checker: ^1.0.0+1
 
 
   # The following adds the Cupertino Icons font to your application.
