@@ -1,7 +1,9 @@
+import 'package:advanced_flutter_clean_architecture/app/constants.dart';
+
 extension NonNullString on String? {
   String orEmpty() {
     if (this == null) {
-      return "";
+      return Constants.empty;
     } else {
       return this!;
     }
@@ -11,7 +13,7 @@ extension NonNullString on String? {
 extension NonNullInteger on int? {
   int orZero() {
     if (this == null) {
-      return 0;
+      return Constants.zero;
     } else {
       return this!;
     }
